Pull `compile` back out of `Utilities`
